+++ PDFDocEx.hh
@@ -1,5 +1,6 @@
 #pragma once
 #include <PDFDoc.h>
+#include <Page.h>
 #include <Zoox.h>
 #include <memory>
 
@@ -13,6 +14,8 @@ public:
     bool isIncremental();
     bool isTagged();
     int getAdbeExtensionLevel();
+    int getNumFontlessPages();
+    int getNumPagesWithImages();
     GString* getExtensions();
     GString* getMetadataString(const char* key);
     GString* getMetadataDateTime(const char* key);
@@ -23,6 +26,7 @@ public:
 private:
     static bool getElemOrAttrData(const ZxElement* elem, const char* nodeName, GString& value, const char* prefix);
     static const char* findXmpPrefix(const ZxElement* elem, const char* nsURI);
+    static bool pageContentIsEmpty(Page* page);
     GString* getXmpValue(const char* nsURI, const char* key, const char* arrayType);
     void getExtensionValues(Object* objExt, GString& data);
     bool openXMP();